Max Phase cleanup (#54)
Bench: 8087895

Simplification of calculation of the max phase. It was originally a method in order to support
potential tuning of this value, but that has been abandoned. Replacing this value with a constant
has led to a speedup.
```
ELO   | 4.36 +- 4.98 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=8MB
LLR   | 2.95 (-2.94, 2.94) [-4.00, 1.00]
Games | N: 8520 W: 2003 L: 1896 D: 4621
```
https://chess.honnold.me/test/269/