Prefetch before NN Accumulator updates (#359)
Bench: 3904316

Prefetch is most helpful when executed as soon as possible. In the case of making a move, this is right after the final zobrist is calculated. As this is purely a speed-up patch, I did not run an LTC.

STC

ELO   | 9.37 +- 5.43 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=8MB
LLR   | 2.98 (-2.94, 2.94) [0.00, 4.00]
GAMES | N: 7416 W: 1853 L: 1653 D: 3910