Update perft.c
I applied new changes.

The goal of `.perft()` is two-fold. First, it's used to ensure the accuracy of the move generator by computing node counts from well known [perft positions](https://www.chessprogramming.org/Perft_Results).  Second, it's used to measure the performance of the move generator as well as the `.move()` and `.undo()` functions.