Fix FRC castling when king doesn't move
Bench: 2542136

Fixes #526

If the king doesn't move while castling, then UndoMove would set the
return square to NO_PIECE. This would result in Berserk thinking it
could play mate in 1s with it's queen as the king no longer existed and
couldn't recapture to get out of check (obviously it can).

This issue escaped Perft checks as board->squares was the array impacted
and that's only used by legality checking (not perft).

Untested as this is just an outright bug.